# cmd/core/internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "quebec core 模块后端",
        "title": "quebec core Swagger API接口文档",
        "contact": {
            "name": "Lyon Mu",
            "url": "https://github.com/lyonmu",
            "email": "lyonmu@foxmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/lyonmu/quebec/blob/master/LICENSE"
        },
        "version": "v0.0.1"
    },
    "host": "http://localhost:59024",
    "basePath": "/core/api",
    "paths": {
...
        "/v1/system/role": {
            "post": {
                "description": "添加角色信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "添加角色信息",
                "parameters": [
                    {
                        "description": "角色信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SystemRoleAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "50000,success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/system/role/enable/{id}": {
            "put": {
                "description": "启停角色状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "启停角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态 [1: 启用, 2: 禁用]",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EnableReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "50000,success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/system/role/list": {
            "get": {
                "description": "获取全部角色列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "全部角色列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2
                        ],
                        "type": "integer",
                        "format": "int32",
                        "x-enum-varnames": [
                            "Yes",
                            "No"
                        ],
                        "description": "角色状态 [1: 启用, 2: 禁用]",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "50000,success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/response.SystemRoleListResp"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/system/role/page": {
            "get": {
                "description": "获取角色分页列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "角色分页列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 1000,
                        "minimum": 10,
                        "type": "integer",
                        "default": 10,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            1,
                            2
                        ],
                        "type": "integer",
                        "format": "int32",
                        "x-enum-varnames": [
                            "Yes",
                            "No"
                        ],
                        "description": "角色状态 [1: 启用, 2: 禁用]",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "50000,success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/response.SystemRoleListResp"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/system/role/{id}": {
            "get": {
                "description": "获取角色详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "获取角色详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "50000,success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/response.SystemRoleResp"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "编辑角色信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "编辑角色信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "角色信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SystemRoleAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "50000,success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "50000,success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "code.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 50000
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "constant.YesOrNo": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "Yes",
                "No"
            ]
        },
        "request.EnableReq": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "description": "状态 [1: 启用, 2: 禁用]",
                    "maximum": 2,
                    "minimum": 1,
                    "allOf": [
                        {
                            "$ref": "#/definitions/constant.YesOrNo"
                        }
                    ]
                }
            }
        },
...
        "request.SystemRoleAddReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "remark": {
                    "description": "角色备注",
                    "type": "string"
                }
            }
        },
...
        "response.SystemRoleListResp": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "角色列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.SystemRoleResp"
                    }
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
        "response.SystemRoleResp": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "角色ID",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "remark": {
                    "description": "角色备注",
                    "type": "string"
                },
                "status": {
                    "description": "角色状态 [1: 启用, 2: 禁用]",
                    "allOf": [
                        {
                            "$ref": "#/definitions/constant.YesOrNo"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "token认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}